Added page to search for youtube videos to download.
Updated client to be able to search for videos using key terms. The number of results
may also be specified. Added svelte-awesome library and general component for IconButton
and SearchBar.

Updated server to have an api route for searching for youtube videos. Added YoutubeDL
library to do this. Note: This may make more sense to be part of the client if possible.