Update downloads table and create icon badges.
Updated the table to have one column for status of the video.
Created badge for navbar icons and added icon to display the number
of current videos in the downloads. Created truncate util to
use in places where we need to possibly truncate. Create formatSeconds
util to format seconds to HH:MM:SS.