Added ability to download audio from youtube videos.
The user can now add and remove youtube videos they want to download
the audio for. They are also able to get the resulted audio file from
the website. Websockets to track if the user is connected have been removed
as they cause other issues. This will need to be fixed by implemented a
sort of cleanup script or some sort of polling.