Added notifications store and better error handling.
Added store to provide the interface with notifications to the user.
The user is now notified when a download is complete. There is also
a notification if there are issues when downloading a file. The server
now has better error handling when trying to download files and will
recreate a session if it was not found.